% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to-utf8.R
\name{to_utf8}
\alias{to_utf8}
\title{Deep conversion to UTF-8}
\usage{
to_utf8(x, ...)
}
\arguments{
\item{x}{A character vector.}
}
\description{
Converts all characters directly or indirectly contained in an object to
UTF-8, using the \code{\link{utf8}} class where possible.
}
\examples{
to_utf8(letters)
to_utf8(iris)
}

