% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{transform_lines}
\alias{transform_lines}
\title{Transform a text file}
\usage{
transform_lines(path, fun, file_encoding = "UTF-8", ok = TRUE,
  skipNul = FALSE, sep = "\\n", write_back = TRUE)
}
\arguments{
\item{path}{A vector of file paths.}

\item{fun}{A function that returns a character vector.}

\item{file_encoding}{The encoding to assume for the input file.}

\item{ok}{logical.  Is it OK to reach the end of the connection before
    \code{n > 0} lines are read?  If not, an error will be generated.}

\item{skipNul}{logical: should nuls be skipped?}

\item{sep}{character string.  A string to be written to the connection
    after each line of text.}

\item{write_back}{Should the results of the transformation be written back
to the file?}
}
\value{
A named logical that indicates if the file has changed, NA if an
error occurred
}
\description{
Reads a file from disk, applies a function on the contents, and optionally
writes the file back if different.
}
\seealso{
Other file functions: \code{\link{read_lines_enc}},
  \code{\link{write_lines_enc}}
}
