% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utf8.R
\docType{methods}
\name{utf8}
\alias{Encoding<-,utf8-method}
\alias{as.character.utf8}
\alias{as.data.frame.utf8}
\alias{as.utf8}
\alias{as.utf8.character}
\alias{as.utf8.default}
\alias{as.utf8.utf8}
\alias{format.utf8}
\alias{is.utf8}
\alias{print.utf8}
\alias{utf8}
\title{A simple class for storing UTF-8 strings}
\usage{
utf8(x = NULL)

is.utf8(x)

as.utf8(x, ...)

\method{as.utf8}{default}(x, ...)

\method{as.utf8}{character}(x, ...)

\method{as.utf8}{utf8}(x, ...)

\method{as.character}{utf8}(x, ...)

\method{as.data.frame}{utf8}(x, row.names = NULL, optional = FALSE, ...,
  nm = paste(deparse(substitute(x), width.cutoff = 500L), collapse = " "))

\S4method{Encoding}{utf8}(x) <- value

\method{format}{utf8}(x, ...)

\method{print}{utf8}(x, ...)
}
\arguments{
\item{x}{A vector}

\item{...}{Arguments passed on to further methods.}

\item{row.names}{\code{NULL} or a character vector giving the row
    names for the data frame.  Missing values are not allowed.}

\item{optional}{logical. If \code{TRUE}, setting row names and
    converting column names (to syntactic names: see
    \code{\link{make.names}}) is optional.  Note that all of \R's
    \pkg{base} package \code{as.data.frame()} methods use
    \code{optional} only for column names treatment, basically with the
    meaning of \code{\link{data.frame}(*, check.names = !optional)}.}

\item{nm}{Name of column in new data frame}

\item{value}{Ignored, values other than \code{"UTF-8"} raise a warning.}
}
\description{
The values are stored as a \code{\link{character}} vector.
Special care is taken to make sure that all elements are UTF-8 or ASCII
strings.
}
\examples{
utf8(letters)
utf8("ä")
utf8(iconv("ä", to = "latin1"))
}

