% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_lines_enc}
\alias{write_lines_enc}
\title{Writes to a text file}
\usage{
write_lines_enc(text, path, file_encoding = "UTF-8", sep = "\\n")
}
\arguments{
\item{text}{A character vector}

\item{path}{Path to the file}

\item{file_encoding}{The encoding for the output file.}

\item{sep}{character string.  A string to be written to the connection
    after each line of text.}
}
\description{
This function is a drop-in replacement for \code{\link[=writeLines]{writeLines()}} from disk files.
It always expects text in the UTF-8 encoding, and by default writes in the
UTF-8 encoding with Unix line separators.
}
\seealso{
\code{\link[readr:write_lines]{readr::write_lines()}} for a faster alternative.

Other file functions: 
\code{\link{read_lines_enc}()},
\code{\link{transform_lines_enc}()}
}
\concept{file functions}
