% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enc-package.R
\docType{package}
\name{enc-package}
\title{enc: Portable Tools for 'UTF-8' Character Data}
\description{
Implements an S3 class for storing 'UTF-8' strings, based on regular character vectors.
Also contains routines to portably read and write 'UTF-8' encoded text files,
to convert all strings in an object to 'UTF-8',
and to create character vectors with various encodings.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/krlmlr/enc}
  \item Report bugs at \url{https://github.com/krlmlr/enc/issues}
}

}
\author{
\strong{Maintainer}: Kirill Müller \email{krlmlr+r@mailbox.org}

}
