% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to-utf8-encoding.R
\name{to_encoding}
\alias{to_encoding}
\alias{to_encoding.character}
\alias{to_utf8}
\title{Deep conversion to an encoding}
\usage{
to_encoding(x, ...)

to_utf8(x, ...)

\method{to_encoding}{character}(x, ..., converter)
}
\arguments{
\item{x}{A character vector.}

\item{...}{passed on to methods}

\item{converter}{A function that accepts a character value as first argument}
}
\description{
Converts all characters directly or indirectly contained in an object to
a specific encoding. \code{to_utf8} converts to
UTF-8, using the \code{\link{utf8}} class where possible.
}
\examples{
to_utf8(letters)
to_utf8(iris)
}

